Using wrapResponse
