+++ graph/resolvers/root_mutation.js
@@ -1,32 +1,5 @@
-const {Error: {ValidationError}} = require('mongoose');
-const errors = require('../../errors');
+const wrapResponse = require('../helpers/response');
 const CommentsService = require('../../services/comments');
-
-/**
- * Wraps up a promise to return an object with the resolution of the promise
- * keyed at `key` or an error caught at `errors`.
- */
-const wrapResponse = (key) => (promise) => {
-  return promise.then((value) => {
-    let res = {};
-    if (key) {
-      res[key] = value;
-    }
-    return res;
-  }).catch((err) => {
-    if (err instanceof errors.APIError) {
-      return {
-        errors: [err]
-      };
-    } else if (err instanceof ValidationError) {
-
-      // TODO: wrap this with one of our internal errors.
-      throw err;
-    }
-
-    throw err;
-  });
-};
 
 const RootMutation = {
   createComment(_, {asset_id, parent_id, body}, {mutators: {Comment}}) {
