+++ plugins/coral-plugin-respect/client/components/Icon.js
@@ -1,5 +1,5 @@
 import React from 'react';
 
 export default () => (
-  <i className="fa fa-handshake-o" aria-hidden="true"></i>
+  <i className="fa fa-handshake-o" aria-hidden="true"/>
 );
