+++ plugins/coral-plugin-respect/index.js
@@ -1,11 +1,12 @@
 const typeDefs = require('./server/typeDefs');
+const wrapResponse = require('../../graph/helpers/response');
 
 module.exports = {
   typeDefs,
   resolvers: {
     RootMutation: {
       createRespect(_, {respect: {item_id, item_type}}, {mutators: {Action}}) {
-        return Action.create({item_id, item_type, action_type: 'RESPECT'});
+        return wrapResponse(Action.create({item_id, item_type, action_type: 'RESPECT'}));
       }
     }
   },
